third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/quirks/table-cell-nowrap-minimum-width-calculation-expected.txt

This is a testharness.js-based test.
FAIL The table cell nowrap minimum width calculation quirk, basic assert_equals: quirks mode expected "10px" but got "8px"
Harness: the test ran to completion.
